Converted Take page into a server-side component, made a new TakeComp componet, Fetched Requests From the Database
